Refactor code from component private methods for Windows statistics (memory / cpu) to WinStats class with tests.
